Switch back to Allegro file routines
This is the only way I can get >4GB file support on Windows without
depending on Windows-only APIs.  LLP64 was the dumbest idea ever.

+++ src/spherefs.h
@@ -26,8 +26,8 @@ int         sfs_fputs  (const char* string, sfs_file_t* file);
 size_t      sfs_fread  (void* buf, size_t size, size_t count, sfs_file_t* file);
 void*       sfs_fslurp (sandbox_t* fs, const char* filename, const char* base_dir, size_t *out_size);
 size_t      sfs_fwrite (const void* buf, size_t size, size_t count, sfs_file_t* file);
-bool        sfs_fseek  (sfs_file_t* file, long offset, sfs_seek_t origin);
-long        sfs_ftell  (sfs_file_t* file);
+bool        sfs_fseek  (sfs_file_t* file, long long offset, sfs_seek_t origin);
+long long   sfs_ftell  (sfs_file_t* file);
 
 enum sfs_seek
 {
